+++ src/main/java/com/picky/domain/book/web/dto/BookDTO.java
@@ -1,7 +1,7 @@
 package com.picky.domain.book.web.dto;
 
+import com.picky.domain.book.entity.Book;
 import io.swagger.v3.oas.annotations.media.Schema;
-import java.time.LocalDateTime;
 import java.util.Collections;
 import java.util.List;
 import lombok.Getter;
@@ -40,9 +40,12 @@ public class BookDTO {
     public static BookDTO fromEntity(Book book) {
         return BookDTO.builder()
                 .title(book.getTitle())
-                .author(Collections.singletonList(book.getAuthor()))
-                .coverImage(book.getThumbnail())
+                .authors(Collections.singletonList(book.getAuthor()))
+                .publisher(book.getPublisher())
+                .coverImage(book.getCoverImage())
                 .isbn(book.getIsbn())
+                .publishedAt(String.valueOf(book.getPublishedAt()))
+                .pageCount(book.getPageCount())
                 .build();
     }
 }
